Refactor: #256 ui 모듈의 PushNotification 모델을 domain 모듈로 이동
