+++ app/src/main/java/com/konkuk/medicarecall/domain/model/UserInfo.kt
@@ -1,7 +1,6 @@
 package com.konkuk.medicarecall.domain.model
 
 import com.konkuk.medicarecall.domain.model.type.GenderType
-import com.konkuk.medicarecall.ui.model.PushNotification
 
 data class UserInfo(
     val name: String = "",
